NEW: Replace wheel with wheel motor and add friction
Instead of simulating a wheel we apply force to we simulate a motor
with a wheel we apply voltage to. This better resembles a real world
sumobot. The programmer will later on be able to apply a voltage/PWM
signal to the wheel motor.

The model is a simplified ideal version of a DC motor. There are two
constants which can be adjusted to change the motor behaviour.

Simulate friction by connecting all wheel motors to a frictionJoint and
make its force depend on the normal force (weight of the sumobot). The
friction can be adjusted with a friction coefficient

Also
* Incorrect terminology, change "weight" to "mass"
* Remove unused code for turning wheels
* Add some code for testing a single wheel