NEW: Range sensor
Add a range sensor class to simulate a range IR sensor using
box2d ray cast.

The range sensor's position is relative to a given body, therefore,
it has to be updated together with that body's orientation
and position.

The sensor can be configured with different max distance (min distance
not yet implemented), position and orientation. It returns the
range to the closest object in front of it. Note, currently it also
detects sensor objects, I will fix this in a future commit.

Add five range sensors to the 4-wheel sumobot as test.