NEW: Line detectors
In a real sumobot line detectors are used to detect the white border
on the dohyo. Simulate line detectors with a tiny circular body and
use Box2D contact listener to detect collision with dohyo border. Use
the body user data to know which bodies have collided. For now,
attach four line detectors (with weldJoints), one in each corner of
the sumobot, an debug draw a circle when collision (detection) occurs.

Also make the range sensors only detect collidable objects.