ENH: Refactor into a standalone project
Move the code out of the testbed and into it's own cmake project.
Remove any testbed specific code and replace it with a simple
main.cpp file. Add Box2D as a subproject and build/link it via CMake.
Also upgrade the Box2D version and adapt to the simulator code
to the new Box2D changes.

No graphics yet, so currently the program is just logic.