NEW: Introduce scenes and scene objects
Just like in a typical game engine, the simulator will have
of a scene of objects. Each object is composed of components:
* TransformComponent
    - Position/Size/Orientation
    - The common view of the object between rendering and physics
* RenderableComponent
    - Specifies which render call to use
* PhysicsComponent
    - Will hold the Box2D object (e.g. rigid body)

For now, there is no list of components in an object, instead the
three components are represensted as separate member variables.
Nor are they stored/accessed in a cache-friendly way because the
simulator won't be that performance heavy, but if need arises, I
will extend it to more mimick a real entity component system.

Also add a scene menu to display the available scenes and to enable
switching between them.