NEW: Offload the microcontroller loop to a separate thread
To prevent the microcontroller loop from "hogging" the simulator loop
and potentially degrade the FPS, offload the loop to a separate thread.
Decouple the access to voltage lines between the simulator objects and
microcontroller by letting them have their own arrays, and create a third
(shared) array, protected with a mutex, to synchronize between them.

Rework the voltage line data structure to fit the new design.

Add virtual destructors to the component classes otherwise polymorphic deletion
won't work (deleting the base class calls the subclass' destructor). This
bug became apparent now because the destructor of the microcontroller class
is used to stop the microcontroller's update thread.

Moreover, ensure the static quad array is used when no texCoords are supplied
to drawQuad.