NEW: Textures for sumobot, wheel motor and dohyo
Add nice-looking textures to sumobot app objects. Plain .png
files for the sumobot and dohyo, and .png with sprite sheet for
the wheel motor (for animation). Extend the app objects with
TextureType enums to support different textures per object.

Add direction to spriteAnimation so that spritesheets can be
animated in the opposite direction (index counting down) as well.
Moreover, support stopping the animation because the wheels should
only animate while moving.

Also:
* Fix incorrect start position calculation of the wheels of the
  TopViewSumobot4Wheel.
* Make camera reset when creating a scene, and don't pixel scale
  its position.
* Add a new background and a second sumobot to the sumobot4wheel
  test scene.