NEW: Scalebar and circular sumobot
Add scalebar at the bottom for visualizing the dimensions
at different zoom levels. Represent it as a set of textures
and replace the texture based on the zoom level.

Add two new sumobot blueprints, and properly add support
for circular-body sumobots. Ensure quadcomponent can take
a circle transform so that a circular body can still be
rendered as a quad texture.