FIX: Make the vehicle physics behave better
The vehicles don't behave as expected when applying unequal forces
(voltages) to each wheel motor. They drive straight when they should
go in a crooked path. There is also a "backlash" when rotating and
then letting go (the vehicle stops rotatin and then starts rotating
in the opposite direction by itself). The torque friction from the
b2FrictionJoint is the main root cause of these issues.

Remove the torqueFrictionCoefficient. Also, remove the friction
from the vehicle body because it's not realistic for it to create
a friction when it's not touching the ground. Instead, "load" the mass
of the body onto the wheels instead. Thus, all friction (sideway + coloumb)
act on the wheels now resulting in more realistic physics.

The friction could, of course, be improved further, but hopefully this
is good enough for now.

And some minor things:
* Make sure the frictionCoefficient can be set inside the robot
specification struct.
* Reduce the minimum speed at which the wheel animation runs