ENH: Separate the physics and rendering update rate
Now the physics and rendering update rate are the same (limited by
VSync). This is an unnecessary limitation and further limits the
update rate of the controller code, which becomes apparent when the
sensor detection is too slow.

Instead of having the physics be updated once before rendering, have
it update as many times at it can according to a fixed update rate.
The approach is the same as laid out by
https://gafferongames.com/post/fix_your_timestep/. This is also better
because now the physics step time is actually fixed, which Box2D likes.

With this change the physics rate can run up to 1000 Hz or faster depending
on the speed of the computer. And note if the configured step time is too
small, the simulation breaks, so introduce a warning for that. Also, if
the physics update rate is close to that of the rendering rate, the rendered
movement will be jittery. The physics update rate needs to be at least 2x
the rendering rate to avoid that and introduce yet another warning for it.
This could also be mitigated by doing graphics interpolation, but would
require quite a bit of work.

Note, the logic (onFixedUpdate callbacks and controllers) are updated at
the same rate as the physics.

Also:
* Make the application base class owner of the scene menu
    - Makes more sense because all applications should have a scene menu
* Add new status information to scene menu: warning and physics update rate
    - And only update it every second to avoid flickering rendering
* Allow the physics step time to be set per scene
    - Depending on the complexity of the scene, some scenes can run faster
* Make the microcontroller update rate settable
    - Now when the physics rate is settable it makes sense to have this settable too
* Safer start of microcontroller
    - Don't start the microcontroller loop until physics is started and 500 ms
      has passed
    - Otherwise the physics bugs out
* Remove unneeded includes of Scene.h in SceneObjects
    - To avoid unneeded rebuilds when changing Scene.h
    - Forward declare is enough
* Remove the stepTime argument from all onFixedUpdate functions
    - It's confusing and no one uses it