NEW: Custom ImGui menus
At the moment, we only have a single hard-coded scene menu. It would be
useful to create custom menus per scene to more easily tweak the scene
objects (e.g. tweaking parameters) and to show statistics.

Add a new class ImGuiMenu to make this possible.

Also, when a scene object is destroyed (without destroying the scene),
it should be removed from the scene object list.