ENH: Microcontroller sleep in physics step instead of OS time
At the moment, we use the sleep function of the operating system to
sleep in the microcontroller code. This is unreliable, because there
is no exact guarantee for how long a thread will sleep, it's determined
by the OS scheduler. On top of that, the frame time and the physics steps
taken vary. This leads to the simulation taking a different number of
physics steps each call to sleep, which makes the makes the entire
simulation very inconsistent.

To solve this, create a new sleep function that sleeps for X milliseconds
counted in physics steps instead of real-time. For example, if we
request to sleep for 100 ms and the physics step time is 1 ms, we
block for 100 physics steps. This makes the sleep time completely
deterministc and in turn the simulation more consistent.

Implement the sleep function in the Microcontroller class and share it
with the C code as a callback function similar to set/get_voltage.

Moreover, we use the OS sleep function to sleep each execution of
the microcontrollor loop function. This is another thing that makes
the simulation inconsistent. The parameter updateRateHz is also misleading,
because it's not only determined by the sleep time but also by how long the
loop function takes to execute. So, remove this sleep, and leave it up
to the C code to call the new sleep function instead.

As a bonus, this enables us to change the simulation speed (slow it
down or speed it up). It's just a matter of multiplying the physics
step time with a multiplier. I will add it in a future commit.