ENH: Make microcontroller C code cleaner and fix initial lag
Currently, the microcontroller C code has to define a setup function
and a loop function. The setup function is called once and the loop
function is called repeatedly. This differs from the code of a real
microcontroller (or C program) where everything starts with a single
main function.

The reason we have a loop function is because the thread must regularly
check for a "is_running" condition so that we can terminate it when the
microcontroller class is destroyed.

But, there is another way, where we can allow for an endless main function and
still let the thread be terminated. It's inspired by how boost does it. We
wrap the main function in a try-catch, and let the get/set_voltage and sleep_ms
function check the running condition and throw an exception if it's false.
Of course, for this to work, the main function must call these functions
regularly, but it almost certainly will, because there is not much use
for it if it doesn't.

Do some additional restructuring of the C files to make it even easier to
create a new microcontroller C binding. The main.c file will now only have
to include a header for the set/get_voltage/sleep functions and define
_main function. Hard to see how it could get any leaner.

Fix the initial lag when a new scene is created. The problem is that the
rendering spikes when a new scene is created, which causes us to initially
take many more physics steps. Work around it by skipping the physics for
the first couple of frames. Also, with this fix in place, remove the startup
sleep time from the microcontroller as we no longer need it.