NEW: Tuning menu for sumobot
Similar to the physics bot, add a tuning menu for the sumobot
to tweak mass, motor paramters and to see speed/acceleration
statistics. Currently, the menu is defined inside the
sumobottestcene, but will be moved to a separate class in
a future commit.

Also, rename GenericBody to SimpleBotBody for clarification.