FIX: Only one C-microcontroller can run at a time
I overlooked an obvious issue. The C functions are not part of
an instantiated class, so we can re-use them the same way as C++
functions. That's why we are passing userdata in the first place...

At the moment, we have a single variable to hold the userdata,
which gets overwritten every time we instanitate a new C microcontroller.
This means, all the microcontroller threads will call the functions
of the same object, which further means it won't work if we try to
run multiple C microcontrollers at the same time.

To fix this, add alternative functions that explicitly take userdata
as an argument. These must be used if the microcontroller should have
several simulated instances of itself. But, still keep the old functions,
because passing userdata is cumbersome, and makes the code uglier.
And we can still simulate a single C microcontroller without passing
userdata among other C microcontrollers that do. Of course, another way
is to simulate a single C microcontroller and multiple C++ microcontrollers.

The same goes for the state variables inside the C file. They must be
defined inside the main function, otherwise the microcontrollers will
try to write to the same variable. So, in the current example, move the
currentState definition into the main function.

There is also no need for the setup functions, instead wrap the Microcontroller
class declaration in ifdef and the C callbacks in extern "C", and include
them right into the microcontroller_c_functions.c. Replace the setup
functions with a single set_userdata function.