ENH: Make motors stoppable and rework sumobot C code
Stop the motors when they go out of the dohyo, which is a
more realistic behaviour because most dohyos are a few
centimeters above the ground. Neither does it make any
sense for the sumobot to drive outside the ring as then
it has already lost.

Rework the C microcontroller example for the sumobot.
Make it better at retreating from the line, change
applied motor duty cycles, add enemy detection, and
more.

Prefix the C-callbacks with _cb instead, to distinguish them.